--- encode/gif_test.go
@@ -1,14 +1,0 @@
-package encode
-
-import (
-	"testing"
-)
-
-func TestGIFEncoder(t *testing.T) {
-
-	err := testEncoder("test.gif")
-
-	if err != nil {
-		t.Fatal(err)
-	}
-}
