--- encode/jpeg_test.go
@@ -1,14 +1,0 @@
-package encode
-
-import (
-	"testing"
-)
-
-func TestJPEGEncoder(t *testing.T) {
-
-	err := testEncoder("test.jpg")
-
-	if err != nil {
-		t.Fatal(err)
-	}
-}
