+++ encode/tiff.go
